classifier/requirements.txt
django>=3.2.3
djangorestframework>=3.12.4
django-cors-headers>=3.7.0
numpy>=1.19.5
pandas>=1.4.1
python-decouple>=3.6
scikit-learn>=1.0.2